python/unknown/bonechewercon/app/challenge/requirements.txt
selenium
Flask
pyjwt
requests
pycryptodome
cryptography